feat: refact github event
