+++ engine/hooks/github.go
@@ -9,64 +9,77 @@ import (
 )
 
 func (s *Service) generatePayloadFromGithubRequest(t *sdk.TaskExecution, event string) (map[string]interface{}, error) {
-	switch event {
-	case "push":
-		return s.generatePayloadFromGithubPushEvent(t)
-	default:
-		return getAllPayloadMap(t.WebHook.RequestBody)
-	}
-}
-
-func (s *Service) generatePayloadFromGithubPushEvent(t *sdk.TaskExecution) (map[string]interface{}, error) {
-	payload := make(map[string]interface{})
 	projectKey := t.Config["project"].Value
 	workflowName := t.Config["workflow"].Value
 
-	var pushEvent GithubPushEvent
-	if err := json.Unmarshal(t.WebHook.RequestBody, &pushEvent); err != nil {
+	var request GithubWebHookEvent
+	if err := json.Unmarshal(t.WebHook.RequestBody, &request); err != nil {
 		return nil, sdk.WrapError(err, "unable ro read github request: %s", string(t.WebHook.RequestBody))
 	}
-	branch := strings.TrimPrefix(pushEvent.Ref, "refs/heads/")
-	if pushEvent.Deleted {
-		err := s.enqueueBranchDeletion(projectKey, workflowName, branch)
 
-		return nil, sdk.WrapError(err, "cannot enqueue branch deletion")
-	}
-	if err := s.stopBranchDeletionTask(branch); err != nil {
-		log.Error("cannot stop branch deletion task for branch %s : %v", branch, err)
-	}
-	payload["git.author"] = pushEvent.HeadCommit.Author.Username
-	payload["git.author.email"] = pushEvent.HeadCommit.Author.Email
+	payload := make(map[string]interface{})
+	payload[GIT_EVENT] = event
 
-	if !strings.HasPrefix(pushEvent.Ref, "refs/tags/") {
-		payload["git.branch"] = branch
-	} else {
-		payload["git.tag"] = strings.TrimPrefix(pushEvent.Ref, "refs/tags/")
-	}
-	payload["git.hash.before"] = pushEvent.Before
-	payload["git.hash"] = pushEvent.After
-	hashShort := pushEvent.After
-	if len(hashShort) >= 7 {
-		hashShort = hashShort[:7]
-	}
-	payload["git.hash.short"] = hashShort
-	payload["git.repository"] = pushEvent.Repository.FullName
-	payload["cds.triggered_by.username"] = pushEvent.HeadCommit.Author.Username
-	payload["cds.triggered_by.fullname"] = pushEvent.HeadCommit.Author.Name
-	payload["cds.triggered_by.email"] = pushEvent.HeadCommit.Author.Email
+	if request.Ref != "" {
+		branch := strings.TrimPrefix(request.Ref, "refs/heads/")
+		if request.Deleted {
+			err := s.enqueueBranchDeletion(projectKey, workflowName, branch)
 
-	if len(pushEvent.Commits) > 0 {
-		payload["git.message"] = pushEvent.Commits[0].Message
+			return nil, sdk.WrapError(err, "cannot enqueue branch deletion")
+		}
+		if err := s.stopBranchDeletionTask(branch); err != nil {
+			log.Error("cannot stop branch deletion task for branch %s : %v", branch, err)
+		}
+
+		if !strings.HasPrefix(request.Ref, "refs/tags/") {
+			payload[GIT_BRANCH] = branch
+		} else {
+			payload[GIT_TAG] = strings.TrimPrefix(request.Ref, "refs/tags/")
+		}
 	}
-	for i := range pushEvent.Commits {
-		pushEvent.Commits[i].Added = nil
-		pushEvent.Commits[i].Removed = nil
-		pushEvent.Commits[i].Modified = nil
+	if request.Before != "" {
+		payload[GIT_HASH_BEFORE] = request.Before
 	}
-	payloadStr, err := json.Marshal(pushEvent)
-	if err != nil {
-		log.Error("Unable to marshal payload: %v", err)
+	if request.After != "" {
+		payload[GIT_HASH] = request.After
+		hashShort := request.After
+		if len(hashShort) >= 7 {
+			hashShort = hashShort[:7]
+		}
+		payload[GIT_HASH_SHORT] = hashShort
 	}
-	payload["payload"] = string(payloadStr)
+
+	getPayloadFromRepository(payload, request.Repository)
+	getPayloadFromCommit(payload, request.HeadCommit)
+
+	if len(request.Commits) > 0 {
+		payload[GIT_MESSAGE] = request.Commits[0].Message
+	}
+
+	for i := range request.Commits {
+		request.Commits[i].Added = nil
+		request.Commits[i].Removed = nil
+		request.Commits[i].Modified = nil
+	}
+	getPayloadStringVariable(payload, request)
+
 	return payload, nil
+}
+
+func getPayloadFromRepository(payload map[string]interface{}, repo *GithubRepository) {
+	if repo == nil {
+		return
+	}
+	payload[GIT_REPOSITORY] = repo.FullName
+}
+
+func getPayloadFromCommit(payload map[string]interface{}, commit *GithubCommit) {
+	if commit == nil {
+		return
+	}
+	payload[GIT_AUTHOR] = commit.Author.Username
+	payload[GIT_AUTHOR_EMAIL] = commit.Author.Email
+	payload[CDS_TRIGGERED_BY_USERNAME] = commit.Author.Username
+	payload[CDS_TRIGGERED_BY_FULLNAME] = commit.Author.Name
+	payload[CDS_TRIGGERED_BY_EMAIL] = commit.Author.Email
 }
